requirements.txt
Django>=4.0
djangorestframework
openpyxl
psycopg2-binary
